Keep regression coverage independent of other pending fixes
